[FIX] resolved PR review (2)
