+++ src/app/activity/[id]/components/price-button/price-button.component.tsx
@@ -1,8 +1,11 @@
 "use client";
 import { ReactElement, useState } from "react";
-import dynamic from "next/dynamic";
+import { clsx } from "clsx";
 
 import { Activity } from "@/lib/data.type";
+
+import PriceListModalComponent from "@/app/activity/[id]/modals/price-list-modal/price-list-modal.component";
+import ButtonComponent from "@/components/button/button.component";
 
 import { openModal } from "@/utils/modal.utils";
 
@@ -15,26 +18,25 @@ type Props = {
 export default function PriceButtonComponent({
   remainingShowings,
 }: Props): ReactElement {
-  const DynamicModal = dynamic(
-    () =>
-      import(
-        "@/app/activity/[id]/modals/price-list-modal/price-list-modal.component"
-      ),
-  );
-
   const [showModal, setShowModal] = useState<boolean>(false);
 
   return (
-    <div className={styles["price-button"]}>
-      <button
-        disabled={remainingShowings === 0}
+    <div
+      className={clsx(
+        styles["price-button"],
+        remainingShowings === 0 && styles.hide,
+      )}
+    >
+      <ButtonComponent
+        className={styles.button}
+        shape="outline"
+        variant="muted"
         onClick={() => openModal<boolean>(setShowModal, true)}
       >
         لیست قیمت بلیت‌ها
-        {" " + ">"}
-        <div className={styles.underline}></div>
-      </button>
-      {showModal && <DynamicModal setShowModal={setShowModal} />}
+        <span className={styles.arrow}>{" " + ">"}</span>
+      </ButtonComponent>
+      {showModal && <PriceListModalComponent setShowModal={setShowModal} />}
     </div>
   );
 }
